# api/swagger.yaml
swagger: "2.0"
info:
  description: ""
  version: "0.0.1"
  title: "TURING API"
host: "localhost:8080"
basePath: "/v1"
schemes:
  - "http"
paths:
  "/projects/{project_id}/routers":
    get:
      tags: ["router"]
      summary: "List routers belonging to project"
      parameters:
        - in: "path"
          name: "project_id"
          description: "project id of the project to retrieve routers from"
          type: "integer"
          required: true
      responses:
        200:
          description: "OK"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/Router"
        400:
          description: "Invalid project_id"
        500:
          description: "Error retrieving routers"
    post:
      tags: ["router"]
      summary: "Create new router in project"
      parameters:
        - in: "path"
          name: "project_id"
          description: "project id of the project to save router"
          type: "integer"
          required: true
        - in: "body"
          name: "body"
          description: "router configuration to save"
          schema:
            $ref: "#/definitions/RouterConfig"
          required: true
      responses:
        200:
          description: "OK"
          schema:
            $ref: "#/definitions/Router"
        400:
          description: "Invalid project_id or router configuration"
        500:
          description: "Unable to save configuration"
  "/projects/{project_id}/routers/{router_id}":
    get:
      tags: ["router"]
      summary: "Get router belonging to project by ID"
      parameters:
        - in: "path"
          name: "project_id"
          description: "project id of the project to retrieve routers from"
          type: "integer"
          required: true
        - in: "path"
          name: "router_id"
          description: "id of the router to be retrieved"
          type: "integer"
          required: true
      responses:
        200:
          description: "OK"
          schema:
            $ref: "#/definitions/Router"
        400:
          description: "Invalid project_id or router_id"
        404:
          description: "No router found"
    put:
      tags: ["router"]
      summary: "Update router"
      parameters:
        - in: "path"
          name: "project_id"
          description: "project id of the project of the router"
          type: "integer"
          required: true
        - in: "path"
          name: "router_id"
          description: "id of the router to update"
          type: "integer"
          required: true
        - in: "body"
          name: "body"
          description: "router configuration to save"
          schema:
            $ref: "#/definitions/RouterConfig"
          required: true
      responses:
        200:
          description: "OK"
          schema:
            $ref: "#/definitions/Router"
        400:
          description: "Invalid project_id, router_id or router configuration"
        500:
          description: "Unable to save configuration"

    delete:
      tags: ["router"]
      summary: "Delete router"
      parameters:
        - in: "path"
          name: "project_id"
          description: "project id of the project of the router"
          type: "integer"
          required: true
        - in: "path"
          name: "router_id"
          description: "id of the router to delete"
          type: "integer"
          required: true
      responses:
        200:
          description: "OK"
          schema:
            type: "object"
            properties:
              router_id:
                type: "integer"
        400:
          description: "Invalid project_id or router_id"
        404:
          description: "Router not found"
        500:
          description: "Unable to delete router"
  "/projects/{project_id}/routers/{router_id}/deploy":
    post:
      tags: ["router"]
      summary: "Deploy the current router configuration"
      parameters:
        - in: "path"
          name: "project_id"
          description: "id of the project that the router belongs to"
          type: "integer"
          required: true
        - in: "path"
          name: "router_id"
          description: "id of the router to be deployed"
          type: "integer"
          required: true
      responses:
        202:
          description: "Accepted"
          schema:
            type: object
            properties:
              router_id:
                type: "integer"
              version:
                type: "integer"
        400:
          description: "Invalid project_id, router_id or deploy request"
        404:
          description: "No router version found"
  "/projects/{project_id}/routers/{router_id}/undeploy":
    post:
      tags: ["router"]
      summary: "Undeploy router configuration"
      parameters:
        - in: "path"
          name: "project_id"
          description: "id of the project that the router belongs to"
          type: "integer"
          required: true
        - in: "path"
          name: "router_id"
          description: "id of the router to undeploy"
          type: "integer"
          required: true
      responses:
        200:
          description: "OK"
          schema:
            type: object
            properties:
              router_id:
                type: "integer"
        400:
          description: "Invalid project_id or router_id"
        404:
          description: "Router not found"
        500:
          description: "Error undeploying router version"
  "/projects/{project_id}/routers/{router_id}/versions":
    get:
      tags: ["router"]
      summary: "List router config versions"
      parameters:
        - in: "path"
          name: "project_id"
          description: "project id of the project to retrieve routers from"
          type: "integer"
          required: true
        - in: "path"
          name: "router_id"
          description: "id of the router to be retrieved"
          type: "integer"
          required: true
      responses:
        200:
          description: "OK"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/RouterVersion"
        400:
          description: "Invalid project_id or router_id"
        404:
          description: "No router versions found"
  "/projects/{project_id}/routers/{router_id}/versions/{version}":
    get:
      tags: ["router"]
      summary: "Get specific router config version"
      parameters:
        - in: "path"
          name: "project_id"
          description: "project id of the project to retrieve routers from"
          type: "integer"
          required: true
        - in: "path"
          name: "router_id"
          description: "id of the router to be retrieved"
          type: "integer"
          required: true
        - in: "path"
          name: "version"
          description: "version of router configuration to be retrieved"
          type: "integer"
          required: true
      responses:
        200:
          description: "OK"
          schema:
            $ref: "#/definitions/RouterVersion"
        400:
          description: "Invalid project_id, router_id or version_id"
        404:
          description: "No router version found"

    delete:
      tags: ["router"]
      summary: "Delete router version"
      parameters:
        - in: "path"
          name: "project_id"
          description: "project id of the project of the router"
          type: "integer"
          required: true
        - in: "path"
          name: "router_id"
          description: "id of the router"
          type: "integer"
          required: true
        - in: "path"
          name: "version"
          description: "version of router configuration to delete"
          type: "integer"
          required: true
      responses:
        200:
          description: "OK"
          schema:
            type: "object"
            properties:
              router_id:
                type: "integer"
              version:
                type: "integer"
        400:
          description: "Invalid project_id, router_id or version"
        404:
          description: "Router version not found"
        500:
          description: "Unable to delete router version"
  "/projects/{project_id}/routers/{router_id}/versions/{version}/deploy":
    post:
      tags: ["router"]
      summary: "Deploy specified version of router configuration"
      parameters:
        - in: "path"
          name: "project_id"
          description: "id of the project that the router belongs to"
          type: "integer"
          required: true
        - in: "path"
          name: "router_id"
          description: "id of the router to be deployed"
          type: "integer"
          required: true
        - in: "path"
          name: "version"
          description: "version of router configuration to be deployed"
          type: "integer"
          required: true
      responses:
        202:
          description: "Accepted"
          schema:
            type: object
            properties:
              router_id:
                type: "integer"
              version:
                type: "integer"
        400:
          description: "Invalid project_id, router_id, version_id or deploy request"
        404:
          description: "No router version found"
  "/projects/{project_id}/routers/{router_id}/logs":
    get:
      tags: ["logs"]
      summary: "List logs for a router, default to 'router' component type and the currently deployed version"
      parameters:
        - in: "path"
          name: "project_id"
          description: "id of the project that the router belongs to"
          type: "integer"
          required: true
        - in: "path"
          name: "router_id"
          description: "id of the router"
          type: "integer"
          required: true
        - in: "query"
          name: "component_type"
          description: "component type of the router to get the logs from, default to 'router' component type"
          type: "string"
          enum: ["router", "enricher", "ensembler"]
          required: false
        - in: "query"
          name: "container"
          description: "container to get the logs from, default to 'user-container', the default container name in Knative"
          type: "string"
          required: false
        - in: "query"
          name: "since_time"
          description: "RFC 3399 timestamp from which to retrieve the logs e.g. 2002-10-02T15:00:00Z. The logs retrieved will have timestamp after (but not including) since_time"
          type: "string"
          required: false
        - in: "query"
          name: "tail_lines"
          description: "number of lines from the end of the logs to retrieve. If both tail_lines and head_lines are set, tail_lines will be ignored"
          type: "string"
          required: false
        - in: "query"
          name: "head_lines"
          description: "number of lines from the start of the logs to retrieve. If both tail_lines and head_lines are set, tail_lines will be ignored"
          type: "string"
          required: false
        - in: "query"
          name: "previous"
          description: "if true, return the logs from previous terminated container in all pods, default to false"
          type: "boolean"
          required: false
      responses:
        200:
          description: "OK"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/PodLog"
        400:
          description: "Invalid value for the router id or query strings"
        500:
          description: "Failed to list pod logs due to internal error"
  "/projects/{project_id}/routers/{router_id}/versions/{version}/logs":
    get:
      tags: ["logs"]
      summary: "List logs for a router with a specific version, default to 'router' component type"
      parameters:
        - in: "path"
          name: "project_id"
          description: "id of the project that the router belongs to"
          type: "integer"
          required: true
        - in: "path"
          name: "router_id"
          description: "id of the router"
          type: "integer"
          required: true
        - in: "path"
          name: "version"
          description: "router version value to get the logs from"
          type: "string"
          required: true
        - in: "query"
          name: "component_type"
          description: "component type of the router to get the logs from, default to 'router' component type"
          type: "string"
          enum: ["router", "enricher", "ensembler"]
          required: false
        - in: "query"
          name: "container"
          description: "container to get the logs from, default to 'user-container', the default container name in Knative"
          type: "string"
          required: false
        - in: "query"
          name: "since_time"
          description: "RFC 3399 timestamp from which to retrieve the logs e.g. 2002-10-02T15:00:00Z"
          type: "string"
          required: false
        - in: "query"
          name: "tail_lines"
          description: "number of lines from the end of the logs to retrieve, this takes priority over 'since_time'"
          type: "string"
          required: false
        - in: "query"
          name: "previous"
          description: "if true, return the logs from previous terminated container in all pods, default to false"
          type: "boolean"
          required: false
      responses:
        200:
          description: "OK"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/PodLog"
        400:
          description: "Invalid value for the router, version id or query strings"
        500:
          description: "Failed to list pod logs due to internal error"
  "/projects/{project_id}/routers/{router_id}/alerts":
    get:
      tags: ["alerts"]
      summary: "List alerts for a router"
      parameters:
        - in: "path"
          name: "project_id"
          description: "id of the project the router belongs to"
          type: "integer"
          required: true
        - in: "path"
          name: "router_id"
          description: "id of the router for the alerts"
          type: "integer"
          required: true
      responses:
        200:
          description: "OK"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/Alert"
        400:
          description: "Invalid project_id or router_id"
        404:
          description: "project_id or router_id not found"
        500:
          description: "Error retrieving the alerts for the router"
    post:
      tags: ["alerts"]
      summary: "Create a new alert for a router"
      parameters:
        - in: "path"
          name: "project_id"
          description: "id of the project the router belongs to"
          type: "integer"
          required: true
        - in: "path"
          name: "router_id"
          description: "id of the router for the alert"
          type: "integer"
          required: true
        - in: "body"
          name: "body"
          description: "alert specification"
          schema:
            $ref: "#/definitions/Alert"
          required: true
      responses:
        200:
          description: "OK"
          schema:
            $ref: "#/definitions/Alert"
        400:
          description: "Invalid project_id or router_id"
        404:
          description: "project_id or router_id not found"
        500:
          description: "Error creating new alert"
  "/projects/{project_id}/routers/{router_id}/alerts/{alert_id}":
    get:
      tags: ["alerts"]
      summary: "Get the alert for a router by ID"
      parameters:
        - in: "path"
          name: "project_id"
          description: "id of the project the router belongs to"
          type: "integer"
          required: true
        - in: "path"
          name: "router_id"
          description: "id of the router for the alerts"
          type: "integer"
          required: true
        - in: "path"
          name: "alert_id"
          description: "id of the alert"
          type: "integer"
          required: true
      responses:
        200:
          description: "OK"
          schema:
            $ref: "#/definitions/Alert"
        400:
          description: "Invalid project_id, router_id or alert_id"
        404:
          description: "Alert not found"
        500:
          description: "Error retrieving the alert"
    put:
      tags: ["alerts"]
      summary: "Update the alert for a router by ID"
      parameters:
        - in: "path"
          name: "project_id"
          description: "id of the project the router belongs to"
          type: "integer"
          required: true
        - in: "path"
          name: "router_id"
          description: "id of the router for the alerts"
          type: "integer"
          required: true
        - in: "path"
          name: "alert_id"
          description: "id of the alert"
          type: "integer"
          required: true
        - in: "body"
          name: "body"
          description: "alert specification"
          schema:
            $ref: "#/definitions/Alert"
          required: true
      responses:
        200:
          description: "OK"
          schema:
            $ref: "#/definitions/Alert"
        400:
          description: "Invalid project_id, router_id or alert_id"
        404:
          description: "Alert not found"
        500:
          description: "Error updating the alert"
    delete:
      tags: ["alerts"]
      summary: "Delete the alert for a router by ID"
      parameters:
        - in: "path"
          name: "project_id"
          description: "id of the project the router belongs to"
          type: "integer"
          required: true
        - in: "path"
          name: "router_id"
          description: "id of the router for the alerts"
          type: "integer"
          required: true
        - in: "path"
          name: "alert_id"
          description: "id of the alert"
          type: "integer"
          required: true
      responses:
        200:
          description: "Alert deleted successfully"
        400:
          description: "Invalid project_id, router_id or alert_id"
        404:
          description: "Alert not found"
        500:
          description: "Error deleting the alert"
  "/experiment-engines":
    get:
      tags: ["experiments"]
      summary: "List the available experiment engines"
      responses:
        200:
          description: "OK"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/ExperimentEngine"
        500:
          description: "Error querying experiment engines"
  "/experiment-engines/{engine}/clients":
    get:
      tags: ["experiments"]
      summary: "List the clients configured on the given experiment engine"
      parameters:
        - in: "path"
          name: "engine"
          description: "name of the experiment engine"
          type: "string"
          required: true
      responses:
        200:
          description: "OK"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/ExperimentClient"
        400:
          description: "Invalid engine name"
        500:
          description: "Error querying experiment engine clients"
  "/experiment-engines/{engine}/experiments":
    get:
      tags: ["experiments"]
      summary: "List the experiments configured on the given experiment engine"
      parameters:
        - in: "path"
          name: "engine"
          description: "name of the experiment engine"
          type: "string"
          required: true
        - in: "query"
          name: "client_id"
          description: "internal (unique) id of the client to filter experiments by"
          type: "string"
      responses:
        200:
          description: "OK"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/Experiment"
        400:
          description: "Invalid engine name"
        500:
          description: "Error querying experiments for the given experiment engine"
  "/experiment-engines/{engine}/variables":
    get:
      tags: ["experiments"]
      summary: "List the variables configured for the given client and/or experiment(s)"
      parameters:
        - in: "path"
          name: "engine"
          description: "name of the experiment engine"
          type: "string"
          required: true
        - in: "query"
          name: "client_id"
          description: "internal (unique) id of the client."
          type: "string"
        - in: "query"
          name: "experiment_id"
          description: "comma-separated internal (unique) ids of the experiments."
          type: "string"
      responses:
        200:
          description: "OK"
          schema:
            $ref: "#/definitions/ExperimentVariables"
        400:
          description: "Invalid engine name, client_id or experiment_id(s)"
        500:
          description: "Error querying variables for the given client / experiment(s)"
  "/projects/{project_id}/routers/{router_id}/events":
    get:
      tags: ["router"]
      summary: "Get deployment events associated with this router"
      parameters:
        - in: "path"
          name: "project_id"
          description: "id of the project that the router belongs to"
          type: "integer"
          required: true
        - in: "path"
          name: "router_id"
          description: "id of the router to be deployed"
          type: "integer"
          required: true
      responses:
        200:
          description: "Get events"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/Event"
        400:
          description: "Invalid project_id or router_id"
        404:
          description: "No router version found"
securityDefinitions:
  Bearer:
    type: apiKey
    name: Authorization
    in: header
security:
  - Bearer: []
definitions:
  Router:
    type: "object"
    properties:
      id:
        type: "integer"
        readOnly: true
      name:
        type: "string"
      created_at:
        type: "string"
        format: "date-time"
        readOnly: true
      updated_at:
        type: "string"
        format: "date-time"
        readOnly: true
      project:
        type: "integer"
      environment_name:
        type: "string"
      endpoint:
        type: "string"
      status:
        $ref: "#/definitions/RouterStatus"
      config:
        $ref: "#/definitions/RouterVersion"
  RouterVersion:
    type: "object"
    properties:
      id:
        type: "integer"
        readOnly: true
      created_at:
        type: "string"
        format: "date-time"
        readOnly: true
      updated_at:
        type: "string"
        format: "date-time"
        readOnly: true
      router:
        type: "object"
        properties:
          Router:
            type: "object"
            properties:
              id:
                type: "integer"
              name:
                type: "string"
              created_at:
                type: "string"
                format: "date-time"
                readOnly: true
              updated_at:
                type: "string"
                format: "date-time"
                readOnly: true
              project:
                type: "integer"
              environment_name:
                type: "string"
              endpoint:
                type: "string"
              status:
                $ref: "#/definitions/RouterStatus"
      version:
        type: "integer"
      status:
        $ref: "#/definitions/RouterVersionStatus"
      error:
        type: "string"
      image:
        type: "string"
      routes:
        type: "array"
        items:
          $ref: "#/definitions/Route"
      default_route:
        type: "string"
      experiment_engine:
        type: "object"
        properties:
          type:
            $ref: "#/definitions/ExperimentEngineType"
          config:
            $ref: "#/definitions/ExperimentConfigResponse"
      resource_request:
        $ref: "#/definitions/ResourceRequest"
      timeout:
        type: "string"
      log_config:
        type: "object"
        properties:
          log_level:
            $ref: "#/definitions/LogLevel"
          custom_metrics_enabled:
            type: "boolean"
          fiber_debug_log_enabled:
            type: "boolean"
          jaeger_enabled:
            type: "boolean"
          result_logger_type:
            $ref: "#/definitions/ResultLoggerType"
          bigquery_config:
            $ref: "#/definitions/BigQueryConfig"
          kafka_config:
            $ref: "#/definitions/KafkaConfig"
      enricher:
        $ref: "#/definitions/Enricher"
      ensembler:
        $ref: "#/definitions/Ensembler"
  RouterConfig:
    type: "object"
    required:
    - environment_name
    - name
    - config
    properties:
      environment_name:
        type: "string"
      name:
        type: "string"
      config:
        type: "object"
        required:
        - routes
        - default_route_id
        - experiment_engine
        - timeout
        - log_config
        properties:
          routes:
            type: "array"
            items:
              $ref: "#/definitions/Route"
          rules:
            type: "array"
            items:
              $ref: "#/definitions/TrafficRule"
          default_route_id:
            type: "string"
          experiment_engine:
            type: "object"
            properties:
              type:
                $ref: "#/definitions/ExperimentEngineType"
              config:
                $ref: "#/definitions/ExperimentConfigRequest"
          resource_request:
            $ref: "#/definitions/ResourceRequest"
          timeout:
            type: "string"
          log_config:
            type: "object"
            properties:
              result_logger_type:
                $ref: "#/definitions/ResultLoggerType"
              bigquery_config:
                $ref: "#/definitions/BigQueryConfig"
              kafka_config:
                $ref: "#/definitions/KafkaConfig"
          enricher:
            $ref: "#/definitions/Enricher"
          ensembler:
            $ref: "#/definitions/Ensembler"
...
  PodLog:
    type: object
    description: "PodLog object contains a log item from a container in a pod"
    properties:
      timestamp:
        type: "string"
        format: "date-time"
        description: "Log timestamp in RFC3339 format"
        example: "2020-07-10T21:13:51Z"
      environment:
        type: "string"
        description: "Environment name of the router running the container that produces this log"
        example: "staging"
      namespace:
        type: "string"
        description: "Kubernetes namespace where the pod running the container is created"
        example: "cartography"
      pod_name:
        type: "string"
        description: "Pod name running the container that produces this log"
        example: "nearby-turing-router-90-deployment-7b49b4bcb5-jlr7n"
      container_name:
        type: "string"
        description: "Container name that produces this log"
        example: "user-container"
      text_payload:
        type: "string"
        description: "Log in text format, either text_payload or json_payload will be set but not both"
        example: "Example log from a container"
      json_payload:
        type: "object"
        description: "Log in JSON format, either text_payload or json_payload will be set but not both"
        example: {"level":"info","msg":"Initializing Nop Result Logger","router_version":"test-1-1.integration-test","ts":1594386832.6632557}
...
  ExperimentEngine:
    type: "object"
    properties:
      name:
        type: "string"
      client_selection_enabled:
        type: "boolean"
      experiment_selection_enabled:
        type: "boolean"
  ExperimentClient:
    type: "object"
    required:
      - id
      - username
    properties:
      id:
        type: "string"
      username:
        type: "string"
      passkey:
        type: "string"
...
  RouterStatus:
    type: "string"
    enum:
      - "deployed"
      - "undeployed"
      - "failed"
      - "pending"
    default: "pending"
...